refactor(dashboard-rework): Renamed some components to better show what they do
