+++ apps/website/src/components/reward-chart-card.tsx
@@ -1,4 +1,3 @@
-import { getChestCountPerRarity } from "@/actions/rarity";
 import {
   Card,
   CardContent,
@@ -6,29 +5,32 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import ChestCountRarityChart from "./rarity-chart";
 import { Suspense } from "react";
 import { Skeleton } from "./ui/skeleton";
 import { FilterConfig } from "@/types/common";
+import { getChestCountPerReward } from "@/actions/reward";
+import ChestCountRewardChart from "./reward-chart";
 
-interface RarityCardProps {
+interface RewardChartCardProps {
   filters: FilterConfig;
 }
 
-export default async function RarityCard({ filters }: RarityCardProps) {
-  const rarities = await getChestCountPerRarity(filters);
+export default async function RewardChartCard({
+  filters,
+}: RewardChartCardProps) {
+  const rewards = await getChestCountPerReward(filters);
 
   return (
     <Card className="shadow-md">
       <CardHeader>
-        <CardTitle>Rarities</CardTitle>
+        <CardTitle>Rewards</CardTitle>
         <CardDescription>
-          Shows the number of treasure chest opened per rarity
+          Shows the number of treasure chest opened per reward
         </CardDescription>
       </CardHeader>
       <CardContent>
         <Suspense fallback={<Skeleton className="aspect-video" />}>
-          <ChestCountRarityChart rarities={rarities} />
+          <ChestCountRewardChart rewards={rewards} />
         </Suspense>
       </CardContent>
     </Card>
